move LayerStateStack to its own files
